migrate click and change helper functions from individual files directly to App.js
According to colocation principles, click and change helper functions which were used in App.js were migrated from external files directly to App.js. Also their tests were migrated to the App.test.js. Individual external function and test files were removed.
