+++ src/App.js
@@ -9,8 +9,6 @@ import { Chart } from "./components/Chart.js";
 import { Filters } from "./components/Filters.js";
 import { Button } from "./components/Button.js";
 import { parseData } from "./functions/parseData";
-import { handleChangeHelper } from "./functions/handleChangeHelper";
-import { handleClickHelper } from "./functions/handleClickHelper";
 
 const App = () => {
   const [data, setData] = useState([]);
@@ -82,3 +80,40 @@ const App = () => {
   );
 };
 export default App;
+
+export const handleChangeHelper = (event, data, selectedValues) => {
+  // Function is triggered on every change of Filter component. Receives information about event as input.
+  // Sets selectedValues state with information about which columns and their values were selected.
+  // Sets useFilter state to false when new filter was selected but button Apply was not yet clicked.
+  if (data.value === undefined || data.value.length === 0) {
+    if (data.placeholder === "campaign") {
+      const { campaign, ...newArray } = { ...selectedValues };
+      return newArray;
+    } else if (data.placeholder === "datasource") {
+      const { datasource, ...newArray } = { ...selectedValues };
+      return newArray;
+    }
+  } else {
+    return Object.assign(
+      { ...selectedValues },
+      {
+        [data.placeholder]: data.value
+      }
+    );
+  }
+};
+
+export const handleClickHelper = selectedValues => {
+  // Function is triggered on every click on the Button Apply component.
+  // Sets useFilter state to true when the button Apply was clicked (unless filters are empty).
+  if (
+    !(
+      Object.entries(selectedValues).length === 0 &&
+      selectedValues.constructor === Object
+    )
+  ) {
+    return { applyFilter: true, filterValue: selectedValues };
+  } else {
+    return { applyFilter: false, filterValue: 0 };
+  }
+};
